+++ app/memiavl.go
@@ -13,9 +13,9 @@ import (
 )
 
 const (
-	FlagMemIAVL     = "memiavl.enable"
-	FlagAsyncCommit = "memiavl.async-commit"
-	FlagZeroCopy    = "memiavl.zero-copy"
+	FlagMemIAVL           = "memiavl.enable"
+	FlagAsyncCommitBuffer = "memiavl.async-commit-buffer"
+	FlagZeroCopy          = "memiavl.zero-copy"
 )
 
 func SetupMemIAVL(logger log.Logger, homePath string, appOpts servertypes.AppOptions, baseAppOptions []func(*baseapp.BaseApp)) []func(*baseapp.BaseApp) {
@@ -23,7 +23,7 @@ func SetupMemIAVL(logger log.Logger, homePath string, appOpts servertypes.AppOpt
 		// cms must be overridden before the other options, because they may use the cms,
 		// make sure the cms aren't be overridden by the other options later on.
 		cms := rootmulti.NewStore(filepath.Join(homePath, "data", "memiavl.db"), logger)
-		cms.SetAsyncCommit(cast.ToBool(appOpts.Get(FlagAsyncCommit)))
+		cms.SetAsyncCommitBuffer(cast.ToInt(appOpts.Get(FlagAsyncCommitBuffer)))
 		cms.SetZeroCopy(cast.ToBool(appOpts.Get(FlagZeroCopy)))
 		baseAppOptions = append([]func(*baseapp.BaseApp){setCMS(cms)}, baseAppOptions...)
 	}